+++ skforecast/model_selection/__init__.py
@@ -5,8 +5,8 @@ from ._split import (
 from ._validation import (
     backtesting_forecaster,
     backtesting_forecaster_multiseries,
-    backtesting_stats,
-    backtesting_foundation
+    backtesting_foundation,
+    backtesting_stats
 )
 from ._search import (
     grid_search_forecaster,
@@ -15,6 +15,7 @@ from ._search import (
     grid_search_forecaster_multiseries,
     random_search_forecaster_multiseries,
     bayesian_search_forecaster_multiseries,
+    bayesian_search_foundation,
     grid_search_stats,
     random_search_stats
 )
@@ -24,14 +25,15 @@ __all__ = [
     "OneStepAheadFold",
     "backtesting_forecaster",
     "backtesting_forecaster_multiseries",
-    "backtesting_stats",
     "backtesting_foundation",
+    "backtesting_stats",
     "grid_search_forecaster",
     "random_search_forecaster",
     "bayesian_search_forecaster",
     "grid_search_forecaster_multiseries",
     "random_search_forecaster_multiseries",
     "bayesian_search_forecaster_multiseries",
+    "bayesian_search_foundation",
     "grid_search_stats",
     "random_search_stats",
 ]